analyze: refactor: move polonius stuff into new borrowck module
