+++ c2rust-analyze/src/borrowck/atoms.rs
